Prog Bar fix, plus minor lvl.0 changes
For the bar I removed the masking element because that was causing so many problems, instead the progress is controlled by a ProgressFill being Overlaid by a new ProgressOverlay. It is a static texture of the progress bar's "border, with a transparent interior that will show the not-filled Black ProgressBG texture, or the filled green ProgressFill Texture.
Plus made some minor tweaks to level zero that was requested including and extra box by the house.