Updated Bit Prefab for grabbing and animations
Updated the Bit Prefab to work with Grabbing, that can be disabled for a level that chooses not to use it. Minor tweaks to bits movement script to fixes some bugs introduced with the lower number of FixedUpdate calls (though some minor graphical defects still persist (it is good enough)). Next update will include going over the script fixing comments and adding more graphics.