LEVEL ZERO BAYBEEE, is like 90% done
As stated Level Zero is complete beside two facts, implementing the dynamic exit (where first try of level, exit takes you to the MainMenu, but returning exits to PseudoIsland. the second fact is the level is littered with bugs. one is the Null Reference Exception (where unity expects a not null, and sees null) related to the dialogue code. Hopefully I'm just tired and cant see what is in front of my face, as every other level is having no troubles with it I can see. Due to this the rest of the Level_0 script has basically gone untested, things feel unpredictable. Do Not Merge as it will make any new games unable to pass through it.