95% complete Level 0, and other fixes.
The Dialogue issue has been resolved thanks to Dexter. Level 0's conditional exit has been partial implemented, the issue is checking the Level's completion status only updates at the end of the scene even if the level was completed prior, so a new method for checking level completion status is needed. Under OnStart() the  code will change functionality of the Levels Exit based on a Boolean value. The only other part of Level 0 That is left in complete is the lesson dialogue explaining binary numbers. A small edit was also made to PS_Level_1 where the Background Images were too low and as a result the camera looked over the top of the background, this was resolved.