End Credits Implemented and Level_0 is complete
End Credits scene has been added. All problems have been fixed with Level_0. A new Title BG has been added along with some minor tweaks to menu prefabs.  one major change was adding a backing for the slider texts so they could be readable no matter where the menu is. partially added a popup that would ask the player if they would like to reset the level, this wasn't fully added because it would probably require rewiring for every level to implement it. we can removed it later if we would like.